contract_commission
=======================
Commission in contracts
=======================

This module allows to get default commissions for partner in the recurring
invoices generated from contracts.

Installation
============

This module is auto-installed when you have *account_analytic_analysis* and
*sale_commission* modules installed.

Usage
=====

When invoicing contracts, the agents assigned to the partner of the contract
will now be assigned to the generated invoices.