[FIX] sale_commission: Avoid erasing of invoice agents
Some onchanges were introduced in https://github.com/hurrinico/commission/pull/2 for
workarounding the issue https://github.com/odoo/odoo/issues/17618, but they are causing
that sales orders that have been invoiced with the proper commissions propagated from
sales, to empty their values when changing one of the following values:

* Partner
* Journal
* Invoice date
* Payment term

It should be taking into account that there are a lot of extensions to this module that
can put different agents according some conditions on sales order, but not on invoices
(for example, sale_commission_formula with formula only in sales orders, or the new
sale_commission_pricelist, that bases the commission on the pricelist item used, and that's
only applicable on sales orders).

More over, the problem will still appears if you add in other module another onchange.

The solution to properly bypass the issue is to directly not add the field on the invoice
form view. This way, there are no glitches. To compensate it, I have add it a computed
field that show the same information, or even better, as we reduce the number of fields
to 1 instead of 2, and this field can be used for expanding the given information.