[RFR] sale_commission: Use commission mixins + remove agents from sales also
Seeing that the same problem with onchanges can happen on sales orders, it's
better to join all the common logic in sales and invoices in some mixins
for not duplicating code, and apply the same on both.

Some fields are still specific to the sales order or invoice objects, as same as
the views, but it's a good code reduction.

Investigating further, the problem still remains in sales orders even hiding
the one2many field, so a workaround in the write method is added for mitigating
it, although it can't be totally fixed: commision line records not yet written
on the DB will be lost after changing some fields in the sale order (partner,
fiscal position, delivery method...).