[REF] sale_commission: Ease + unify inheritance
Having 2 methods that are called alternatively depending on the path (existing record
or new record) makes the inheritance harder, having to override both methods.

With this refactoring, we ease that, unifying the method and having 2 arguments for
being handled in it (`self` for existing record and `vals` for to be create records).