[FIX] sale_commission_agent_restrict: do not deactivate base rules
Fix that ended up involving a significant refactor of the entire
module. The goal was mostly to keep the same functionality but also
keep the security rule about hiding private addresses for accounts
that are not agents.

The commit adds a new OCA dependency; there's some code cleanup and
refactoring as well, and many more tests.